# swagger.yaml
openapi: 3.0.0
paths:
...
  /demo/random-key:
    get:
      tags:
        - demo
      description: Generates a new Hedera account with a random private key.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HederaAccount'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /demo/push/random-key:
    get:
      tags:
        - demo
      description: Generates a new Hedera account with a random private key.
      security:
        - bearerAuth: []
      summary: Generates a new Hedera account with a random private key.
      responses:
        '202':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notifications:
    get:
      operationId: NotificationsApi_getAllNotifications
      summary: Get all notifications
      description: Returns all notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns notifications and count.
          headers:
            X-Total-Count:
              description: Count of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: &ref_7
        - notifications
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/new:
    get:
      operationId: NotificationsApi_getNewNotifications
      summary: Get new notifications
      description: Returns new notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns new notifications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_7
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/progresses:
    get:
      operationId: NotificationsApi_getProgresses
      summary: Get progresses
      description: Returns progresses.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns progresses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_7
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/read/all:
    post:
      operationId: NotificationsApi_readAll
      summary: Read all notifications
      description: Returns new notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns notifications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_7
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/delete/{notificationId}:
    delete:
      operationId: NotificationsApi_delete
      summary: Delete notifications up to this point
      description: Returns deleted notifications count.
      parameters:
        - name: notificationId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation. Returns deleted notifications count.
          content:
            application/json:
              schema:
                type: number
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_7
      security:
        - bearer: []
        - bearerAuth: []
...
info:
  title: Guardian
  description: >-
    The Guardian is a modular open-source solution that includes best-in-class
    identity management and decentralized ledger technology (DLT) libraries. At
    the heart of the Guardian solution is a sophisticated Policy Workflow Engine
    (PWE) that enables applications to offer a requirements-based tokenization
    implementation.
  version: 2.16.0
  contact:
    name: API developer
    url: https://envisionblockchain.com
    email: info@envisionblockchain.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: /api/v1
    description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InternalServerErrorDTO:
      type: object
      properties:
        code:
          type: number
        message:
          type: string
      required:
        - code
        - message
...
    NotificationDTO:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum:
            - INFO
            - ERROR
            - WARN
            - SUCCESS
        action:
          type: string
          enum:
            - POLICY_CONFIGURATION
            - POLICY_VIEW
            - POLICIES_PAGE
            - SCHEMAS_PAGE
            - TOKENS_PAGE
            - PROFILE_PAGE
        result:
          type: object
        read:
          type: boolean
        old:
          type: boolean
      required:
        - title
        - message
        - type
        - action
        - result
        - read
        - old
    ProgressDTO:
      type: object
      properties:
        action:
          type: string
        message:
          type: string
        progress:
          type: number
        type:
          type: string
          enum:
            - INFO
            - ERROR
            - WARN
            - SUCCESS
        taskId:
          type: string
      required:
        - action
        - message
        - progress
        - type
        - taskId
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    HederaAccount:
      type: object
      required:
        - id
        - key
      properties:
        id:
          type: string
        key:
          type: string
...
    Task:
      type: object
      properties:
        taskId:
          type: string
        expectation:
          type: number
...